# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Barber Hub API — SaaS platform for barbershops in Germany
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /barbershops:
    get:
      operationId: listBarbershops
      tags: [barbershops]
      summary: List all barbershops
      parameters:
        - in: query
          name: city
          schema:
            type: string
        - in: query
          name: ownerId
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of barbershops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barbershop"
    post:
      operationId: createBarbershop
      tags: [barbershops]
      summary: Create a barbershop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarbershopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barbershop"
  /barbershops/{id}:
    get:
      operationId: getBarbershop
      tags: [barbershops]
      summary: Get a barbershop by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barbershop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barbershop"
        "404":
          description: Not found
    patch:
      operationId: updateBarbershop
      tags: [barbershops]
      summary: Update a barbershop
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarbershopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barbershop"
  /barbershops/{id}/stats:
    get:
      operationId: getBarbershopStats
      tags: [barbershops]
      summary: Get statistics for a barbershop
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barbershop statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarbershopStats"

  # ── Barbers ───────────────────────────────────────────────────────────────────
  /barbers:
    get:
      operationId: listBarbers
      tags: [barbers]
      summary: List barbers
      parameters:
        - in: query
          name: barbershopId
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of barbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Barber"
    post:
      operationId: createBarber
      tags: [barbers]
      summary: Create a barber profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
  /barbers/{id}:
    get:
      operationId: getBarber
      tags: [barbers]
      summary: Get a barber by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Barber
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"
        "404":
          description: Not found
    patch:
      operationId: updateBarber
      tags: [barbers]
      summary: Update a barber
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarberUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Barber"

  # ── Services ──────────────────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - in: query
          name: barbershopId
          schema:
            type: ["integer", "null"]
        - in: query
          name: barberId
          schema:
            type: ["integer", "null"]
        - in: query
          name: clientId
          schema:
            type: ["integer", "null"]
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: date
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ── Clients ───────────────────────────────────────────────────────────────────
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - in: query
          name: barbershopId
          schema:
            type: ["integer", "null"]
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

  # ── Reviews ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Barbershop:
      type: object
      required: [id, name, city, address, rating, reviewCount, nationalityFlag, ownerId]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        city:
          type: string
        address:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        coverImage:
          type: ["string", "null"]
        photos:
          type: array
          items:
            type: string
        specialties:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        nationalityFlag:
          type: string
        nationalityLabel:
          type: ["string", "null"]
        isPremium:
          type: boolean
        isVerified:
          type: boolean
        ownerId:
          type: integer
        createdAt:
          type: string
    BarbershopInput:
      type: object
      required: [name, city, address, ownerId]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        coverImage:
          type: string
        specialties:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        nationalityFlag:
          type: string
        nationalityLabel:
          type: string
        isPremium:
          type: boolean
        ownerId:
          type: integer
    BarbershopUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        coverImage:
          type: string
        specialties:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        nationalityFlag:
          type: string
        nationalityLabel:
          type: string
        isPremium:
          type: boolean
    BarbershopStats:
      type: object
      required: [totalRevenue, totalAppointments, totalClients, avgRating, topService]
      properties:
        totalRevenue:
          type: number
        monthlyRevenue:
          type: number
        totalAppointments:
          type: integer
        completedAppointments:
          type: integer
        totalClients:
          type: integer
        newClientsThisMonth:
          type: integer
        avgRating:
          type: number
        topService:
          type: ["string", "null"]
    Barber:
      type: object
      required: [id, name, barbershopId, nationality]
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        barbershopId:
          type: integer
        barbershopName:
          type: ["string", "null"]
        nationality:
          type: string
        nationalityFlag:
          type: string
        specialties:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        yearsExperience:
          type: ["integer", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
    BarberInput:
      type: object
      required: [name, barbershopId, nationality]
      properties:
        name:
          type: string
          minLength: 1
        bio:
          type: string
        barbershopId:
          type: integer
        nationality:
          type: string
        nationalityFlag:
          type: string
        specialties:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        yearsExperience:
          type: integer
        isAvailable:
          type: boolean
    BarberUpdate:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        nationality:
          type: string
        nationalityFlag:
          type: string
        specialties:
          type: array
          items:
            type: string
        languages:
          type: array
          items:
            type: string
        avatarUrl:
          type: string
        yearsExperience:
          type: integer
        isAvailable:
          type: boolean
...
    Appointment:
      type: object
      required: [id, clientId, barberId, serviceId, barbershopId, scheduledAt, status, price]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        barberId:
          type: integer
        barberName:
          type: ["string", "null"]
        serviceId:
          type: integer
        serviceName:
          type: ["string", "null"]
        barbershopId:
          type: integer
        barbershopName:
          type: ["string", "null"]
        scheduledAt:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        price:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [clientId, barberId, serviceId, barbershopId, scheduledAt, price]
      properties:
        clientId:
          type: integer
        barberId:
          type: integer
        serviceId:
          type: integer
        barbershopId:
          type: integer
        scheduledAt:
          type: string
        price:
          type: number
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, no_show]
        scheduledAt:
          type: string
        notes:
          type: string
    Client:
      type: object
      required: [id, name, barbershopId]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        barbershopId:
          type: integer
        birthdate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        preferredBarberId:
          type: ["integer", "null"]
        lastVisit:
          type: ["string", "null"]
        totalVisits:
          type: integer
        totalSpent:
          type: number
        loyaltyPoints:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, barbershopId]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        barbershopId:
          type: integer
        birthdate:
          type: string
        notes:
          type: string
        preferredBarberId:
          type: integer
        avatarUrl:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        birthdate:
          type: string
        notes:
          type: string
        preferredBarberId:
          type: ["integer", "null"]
        loyaltyPoints:
          type: integer
...