# lib/api-zod/src/generated/types/barber.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Barber Hub API — SaaS platform for barbershops in Germany
 * OpenAPI spec version: 0.1.0
 */

export interface Barber {
  id: number;
  name: string;
  /** @nullable */
  bio?: string | null;
  barbershopId: number;
  /** @nullable */
  barbershopName?: string | null;
  nationality: string;
  nationalityFlag?: string;
  specialties?: string[];
  languages?: string[];
  rating?: number;
  reviewCount?: number;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  isAvailable?: boolean;
  createdAt?: string;
}