# lib/api-zod/src/generated/types/barberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Barber Hub API — SaaS platform for barbershops in Germany
 * OpenAPI spec version: 0.1.0
 */

export interface BarberUpdate {
  name?: string;
  bio?: string;
  nationality?: string;
  nationalityFlag?: string;
  specialties?: string[];
  languages?: string[];
  avatarUrl?: string;
  yearsExperience?: number;
  isAvailable?: boolean;
}