# lib/api-zod/src/generated/types/barbershop.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Barber Hub API — SaaS platform for barbershops in Germany
 * OpenAPI spec version: 0.1.0
 */

export interface Barbershop {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  city: string;
  address: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  website?: string | null;
  rating: number;
  reviewCount: number;
  /** @nullable */
  coverImage?: string | null;
  photos?: string[];
  specialties?: string[];
  languages?: string[];
  nationalityFlag: string;
  /** @nullable */
  nationalityLabel?: string | null;
  isPremium?: boolean;
  isVerified?: boolean;
  ownerId: number;
  createdAt?: string;
}